@@ -1,7 +1,7 @@
 package com.py.cheng.nong.xiang.dailyreadjava.model.adapter;
 
 import android.content.Intent;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
