@@ -2,17 +2,11 @@ package com.py.cheng.nong.xiang.dailyreadjava.model.bean;
 
 import java.util.List;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 /**
  * @author xiang
  * @date 2018/3/5
  */
-@Getter
-@Setter
-@NoArgsConstructor
 public class Date {
 
     private String date;
@@ -21,4 +15,30 @@ public class Date {
 
     private List<TopStory> topStories;
 
+    public Date() {
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public void setDate(String date) {
+        this.date = date;
+    }
+
+    public List<Story> getStories() {
+        return stories;
+    }
+
+    public void setStories(List<Story> stories) {
+        this.stories = stories;
+    }
+
+    public List<TopStory> getTopStories() {
+        return topStories;
+    }
+
+    public void setTopStories(List<TopStory> topStories) {
+        this.topStories = topStories;
+    }
 }
