@@ -4,18 +4,11 @@ import com.alibaba.fastjson.annotation.JSONField;
 
 import java.util.List;
 
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
 /**
  * @author xiang
  * @date 2018/3/5
  */
 
-@Getter
-@Setter
-@NoArgsConstructor
 public class Story {
 
     private List<String> images;
@@ -29,4 +22,46 @@ public class Story {
 
     private String title;
 
+    public Story() {
+    }
+
+    public List<String> getImages() {
+        return images;
+    }
+
+    public void setImages(List<String> images) {
+        this.images = images;
+    }
+
+    public int getType() {
+        return type;
+    }
+
+    public void setType(int type) {
+        this.type = type;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getGaPrefix() {
+        return gaPrefix;
+    }
+
+    public void setGaPrefix(String gaPrefix) {
+        this.gaPrefix = gaPrefix;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
 }
